fix: unify typechecking and ast collection
