+++ packages/vitest/src/node/ast-collect.ts
@@ -34,7 +34,7 @@ interface ParsedSuite extends Suite {
   dynamic: boolean
 }
 
-interface LocalCallDefinition {
+export interface LocalCallDefinition {
   start: number
   end: number
   name: string
@@ -44,6 +44,24 @@ interface LocalCallDefinition {
   dynamic: boolean
   concurrent: boolean | undefined
   tags: string[]
+}
+
+export interface FileInformation {
+  file: File
+  filepath: string
+  parsed: string
+  map: any
+  definitions: LocalCallDefinition[]
+}
+
+export interface AstCollectOptions {
+  /**
+   * Override the pool stored on the resulting File task. Required when
+   * collecting typecheck files because the project's `config.pool` is the
+   * user's runtime pool (e.g. `forks`), not the `typescript` pool that the
+   * typecheck spec uses to compute its task id.
+   */
+  pool?: string
 }
 
 const debug = createDebugger('vitest:ast-collect-info')
@@ -287,15 +305,16 @@ function astParseFile(filepath: string, code: string) {
   }
 }
 
-export function createFailedFileTask(project: TestProject, filepath: string, error: Error): File {
+export function createFailedFileTask(project: TestProject, filepath: string, error: Error, options?: AstCollectOptions): File {
   const config = project.serializedConfig
+  const pool = options?.pool ?? config.pool
   const baseFile = createFileTaskOriginal(
     filepath,
     config.root,
     config.name,
-    config.pool,
+    pool,
     undefined,
-    { typecheck: config.pool === 'typescript', __vitest_label__: config.mergeReportsLabel },
+    { typecheck: pool === 'typescript', __vitest_label__: config.mergeReportsLabel },
   )
   const file: ParsedFile = {
     ...baseFile,
@@ -340,16 +359,18 @@ function createFileTask(
   requestMap: any,
   filepath: string,
   fileTags: string[] | undefined,
+  options?: AstCollectOptions,
 ) {
   const { definitions, ast } = astParseFile(testFilepath, code)
   const config = project.serializedConfig
+  const pool = options?.pool ?? config.pool
   const baseFile = createFileTaskOriginal(
     filepath,
     config.root,
     config.name,
-    config.pool,
+    pool,
     undefined,
-    { typecheck: config.pool === 'typescript', __vitest_label__: config.mergeReportsLabel },
+    { typecheck: pool === 'typescript', __vitest_label__: config.mergeReportsLabel },
   )
   const file: ParsedFile = {
     ...baseFile,
@@ -484,31 +505,55 @@ function createFileTask(
       ],
     }
   }
-  return file
+  return { file, definitions }
 }
 
 export async function astCollectTests(
   project: TestProject,
   filepath: string,
 ): Promise<File> {
+  const information = await astCollectFileInformation(project, filepath)
+  return information.file
+}
+
+export async function astCollectFileInformation(
+  project: TestProject,
+  filepath: string,
+  options?: AstCollectOptions,
+): Promise<FileInformation> {
   const request = await transformSSR(project, filepath)
   const testFilepath = relative(project.config.root, filepath)
   if (!request) {
     debug?.('Cannot parse', testFilepath, '(vite didn\'t return anything)')
-    return createFailedFileTask(
-      project,
+    return {
+      file: createFailedFileTask(
+        project,
+        filepath,
+        new Error(`Failed to parse ${testFilepath}. Vite didn't return anything.`),
+        options,
+      ),
       filepath,
-      new Error(`Failed to parse ${testFilepath}. Vite didn't return anything.`),
-    )
+      parsed: '',
+      map: null,
+      definitions: [],
+    }
   }
-  return createFileTask(
+  const { file, definitions } = createFileTask(
     project,
     testFilepath,
     request.code,
     request.map,
     filepath,
     request.fileTags,
+    options,
   )
+  return {
+    file,
+    filepath,
+    parsed: request.code,
+    map: request.map,
+    definitions,
+  }
 }
 
 async function transformSSR(project: TestProject, filepath: string) {
